Add Ava testing suite.
