+++ package.json
@@ -5,13 +5,17 @@
   "main": "index.js",
   "scripts": {
     "babel-node": "babel-node --presets=env",
+    
     "db:migrate:up": "knex migrate:latest --knexfile ./src/db/knexfile.js ",
     "db:migrate:down": "knex migrate:rollback --knexfile ./src/db/knexfile.js ",
     "db:migrate:make": "knex migrate:make --knexfile ./src/db/knexfile.js ",
     "db:seed": "knex seed:run --knexfile ./src/db/knexfile.js",
+    
+    "test--db:migrate:up": "NODE_ENV=test knex migrate:latest --knexfile ./src/db/knexfile.js",
+    
     "start:dev": "nodemon --exec npm run babel-node -- ./src/server/server.js",
     "webdev:start": "webpack-dev-server",
-    "test": "npm run babel-node index.js"
+    "test": "NODE_ENV=test ava"
   },
   "author": "Ian Deans",
   "license": "MIT",
@@ -24,16 +28,39 @@
     "sqlite3": "^3.1.13"
   },
   "devDependencies": {
+    "ava": "^0.23.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
+    "babel-register": "^6.26.0",
+    "eslint": "^4.10.0",
+    "eslint-plugin-react": "^7.4.0",
     "html-webpack-plugin": "^2.30.1",
     "nodemon": "^1.12.1",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "webpack": "^3.8.1",
     "webpack-dev-server": "^2.9.3"
+  },
+  "ava": {
+    "files": [
+      "**/tests/*.test.js",
+      "!**/not-this-file.js"
+    ],
+    "source": [
+      "**/*.{js,jsx}",
+      "!dist/**/*"
+    ],
+    "concurrency": 5,
+    "failFast": true,
+    "failWithoutAssertions": false,
+    "tap": true,
+    "powerAssert": false,
+    "require": [
+      "babel-register"
+    ]
   }
 }
