+++ src/db/db.js
@@ -1,6 +1,4 @@
 let config = require('./knexfile.js')
-// FIXME: Remove hardcoding of development configuration
-// config = process.env.NODE_ENV ? config[process.env.NODE_ENV] : config['development']
-const db = require('knex')(config.development)
+config = process.env.NODE_ENV ? config[process.env.NODE_ENV] : config['development']
 
-export default db
+export default require('knex')(config)
