+++ src/db/utils.js
@@ -1,0 +1,7 @@
+import db from './db'
+
+const utils = {
+  clearTable: tableName => db(tableName).select().del()
+}
+
+export default utils
